Library

Library

Documentation for TextUserInterfaces.jl.

TextUserInterfaces.fieldjust — Constant.

Dictionary defining values of the field justification.

TextUserInterfaces.fieldopts — Constant.

Dictionary defining values of the field options.

TextUserInterfaces.formcmd — Constant.

Dictionary defining values of the form commands.

TextUserInterfaces.menucmd — Constant.

Dictionary defining values of the menu commands.

TextUserInterfaces.Keystroke — Type.

struct Keystorke

Structure that defines a keystroke.

Fields

value: String representing the keystroke.

ktype: Type of the key (:char, :F1, :up, etc.).

alt: true if ALT key was pressed (only valid if ktype != :char).

ctrl: true if CTRL key was pressed (only valid if ktype != :char).

shift: true if SHIFT key was pressed (only valid if ktype != :char).

source

TextUserInterfaces.ACS_ — Method.

function ACS_(s::Symbol)

Return the symbol s of the acs_map. For example, ACS_HLINE can be obtained from ACS_(:HLINE).

TextUserInterfaces.COLS — Method.

function COLS()

Return the number of columns in the root window. It must be called after initscr().

For more information, consult libncurses documentation.

TextUserInterfaces.LINES — Method.

function LINES()

Return the number of lines in the root window. It must be called after initscr().

For more information, consult libncurses documentation.

TextUserInterfaces.attroff — Function.

function attroff(attrs::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.attron — Function.

function attron(attrs::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.box — Function.

function box(win::Ptr{WINDOW}, verch::jlchtype, horch::jlchtype)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.cbreak — Function.

function cbreak()

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.clear — Function.

function clear()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.clear_window — Method.

function clear_window(win::TUI_WINDOW; clear_type = :all)

Clear the window win according the to clearing type in clear_type:

:all: Clears the entire window.

:to_screen_bottom: Clears everything from the cursor position to the bottom of the screen.

:to_eol: Clear everything from the cursor position to the end of line.

TextUserInterfaces.clrtobot — Function.

function clrtobot()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.clrtoeol — Function.

function clrtoeol()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.create_menu — Method.

function create_menu(names::Vector{T1}, descriptions::Union{Nothing,Vector{T2}} = nothing; kwargs...)

Create a menu with the names names and descriptions descriptions. If descriptions is nothing or is omitted, then the menu descriptions will not be shown.

Keywords

format: If nothing, then the default menu format will be used. Otherwise, it must be a tuple with two integers describing respectively the number of rows and columns. (Default = nothing)

mark: The menu mark. (Default = -)

one_value: If true, then the menu cannot have multiple selections. (Default = true)

show_desc: If true, then the menu description will be shown. (Default = true)

row_major: If true, then the menu will be constructed in a row-major ordering.

ignore_case: If true, then the case will be ignored on pattern matching.

show_match: If true, then the cursor will be moved to within the item name while pattern-matching.

non_cyclic: If true, then the menu will be non cyclic.

TextUserInterfaces.create_panel — Function.

function create_panel(win::TUI_WINDOW, prev::Union{Nothing,TUI_PANEL} = nothing, next::Union{Nothing,TUI_PANEL} = nothing)

Create a panel for the window win in which the previous panel is prev and the next panel is next.

TextUserInterfaces.create_panel_chain — Method.

function create_panel_chain(panels::TUI_PANEL...; circular = true)

Create a chain of panels by assigning the prev and next pointers of each one. If the keyword circular is true, then the list will be circular.

Example

create_panel_chain(panel1, panel2, panel3; circular = false)

nothing <--> panel1 <--> panel2 <--> panel3 <--> nothing

create_panel_chain(panel1, panel2, panel3; circular = true)

.---> panel1 <--> panel2 <--> panel3 <--.
|                                       |
|                                       |
'---------------------------------------'

TextUserInterfaces.create_window — Function.

function createwindow([parent::Union{Nothing,TUIWINDOW}, ]nlines::Integer, ncols::Integer, beginy::Integer, beginx::Integer, id::String = ""; border = true)

Create a window inside the parent window parent. If parent is nothing or if it is omitted, then the root window will be used as the parent window. The new window size will be nlines × ncols and the origin will be placed at (begin_y, begin_x) coordinate of the parent window. The window ID id is used to identify the new window in the global window list.

Keyword

border: If true, then the window will have a border. (Default = true)

border_color: Color mask that will be used to print the border. See function ncurses_color. If negative, then the color will not be changed. (Default = -1)

title: The title of the window, which will only be printed if border is true. (Default = "")

title_color: Color mask that will be used to print the title. See function ncurses_color. If negative, then the color will not be changed. (Default = -1)

TextUserInterfaces.create_window_layout — Method.

function create_window_layout(parent::TUI_WINDOW, vert::Vector{T1}, horz::Vector{T2}) where {T1<:Number, T2<:Number}

Create a window layout inside the parent window parent. If it is omitted, then the root window will be used as parent.

The layout dimensions is obtained from the vectors vert and horz. They will be interpreted as a percentage of the total size. For example, if vert = [50,25,25], then the layout will have three lines in each the first will have 50% of the total size, and the second and third 25%. The same applies for the horz vector for the columns size of the layout.

This function return a matrix with the windows references.

TextUserInterfaces.current_item — Function.

function current_item(menu::Ptr{Cvoid})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.current_item — Method.

current_item(menu::TUI_MENU) = current_item(menu.ptr)

Return the current item of the menu menu.

TextUserInterfaces.current_item_desc — Method.

function current_item_desc(menu::TUI_MENU)

Return the item description of the menu menu.

TextUserInterfaces.current_item_name — Method.

function current_item_name(menu::TUI_MENU)

Return the item name of the menu menu.

TextUserInterfaces.curs_set — Function.

function curs_set(visibility::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.curses_version — Method.

function curses_version()

Return the NCurses version in a named tuple with the following fields:

major: Major version.

minor: Minor version.

patch: Patch version.

TextUserInterfaces.del_panel — Function.

function del_panel(panel::Ptr{Cvoid})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.delwin — Function.

function delwin(win::Ptr{WINDOW})

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.derwin — Function.

function derwin(win::Ptr{WINDOW}, nlines::Integer, ncols::Integer, begin_y::Integer, begin_x::Integer)

Return type: Ptr{TextUserInterfaces.WINDOW}

For more information, consult libncurses documentation.

TextUserInterfaces.destroy_all_panels — Method.

function destroy_all_panels()

Destroy all panels managed by the TUI.

TextUserInterfaces.destroy_all_windows — Method.

function destroy_all_windows()

Destroy all windows managed by the TUI.

TextUserInterfaces.destroy_menu — Method.

function destroy_menu(menu::TUI_MENU)

Destroy the menu menu.

TextUserInterfaces.destroy_panel — Method.

function destroy_panel(panel::TUI_PANEL)

Destroy the panel panel.

TextUserInterfaces.destroy_tui — Method.

function destroy_tui()

Destroy the Text User Interface (TUI).

TextUserInterfaces.destroy_window — Method.

function destroy_window(win::TUI_WINDOW)

Destroy the window win.

TextUserInterfaces.doupdate — Function.

function doupdate()

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.endwin — Function.

function endwin()

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.erase — Function.

function erase()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.field_buffer — Function.

function field_buffer(field::Ptr{Cvoid}, buffer::Integer)

Return type: Cstring

For more information, consult libform documentation.

TextUserInterfaces.form_driver — Function.

function form_driver(form::Ptr{Cvoid}, ch::Integer)

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.free_field — Function.

function free_field(field::Ptr{Cvoid})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.free_form — Function.

function free_form(form::Ptr{Cvoid})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.free_item — Function.

function free_item(item::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.free_menu — Function.

function free_menu(menu::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.get_color_pair — Method.

function get_color_pair(foreground::Symbol, background::Symbol)

Return the ID of the color pair (foreground, background), or nothing if the color pair is not initialized.

TextUserInterfaces.getch — Function.

function getch()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.has_colors — Function.

function has_colors()

Return type: UInt8

For more information, consult libncurses documentation.

TextUserInterfaces.hide_panel — Function.

function hide_panel(panel::Ptr{Cvoid})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.hide_panel — Method.

function hide_panel(panel::TUI_PANEL)

Hide the panel panel.

TextUserInterfaces.hline — Function.

function hline(ch::jlchtype, n::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.init_color — Function.

function init_color(color::Integer, r::Integer, g::Integer, b::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.init_color — Method.

function init_color(name::Symbol, r::Int, g::Int, b::Int)

Initialize the color with name name and RGB color r, g, and b. Notice that the range for the last three variables is [0,1000].

If the color is already initialized, then nothing will be changed.

If the color was initialized, then it returns the color ID. Otherwise, it returns -1.

TextUserInterfaces.init_color_pair — Method.

function init_color_pair(foreground::Symbol, background::Symbol)

Initialize the color pair (foreground, background) and return its ID. If the pair already exists, then just the function just returns its ID.

TextUserInterfaces.init_pair — Function.

function init_pair(pair::Integer, f::Integer, b::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.init_tui — Function.

function init_tui(dir::String = "")

Initialize the Text User Interface (TUI). The full-path of the ncurses directory can be specified by dir. If it is empty or omitted, then it will look on the default library directories.

TextUserInterfaces.initscr — Function.

function initscr()

Return type: Ptr{TextUserInterfaces.WINDOW}

For more information, consult libncurses documentation.

TextUserInterfaces.item_description — Function.

function item_description(menu::Ptr{Cvoid})

Return type: Cstring

For more information, consult libmenu documentation.

TextUserInterfaces.item_index — Function.

function item_index(menu::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.item_name — Function.

function item_name(menu::Ptr{Cvoid})

Return type: Cstring

For more information, consult libmenu documentation.

TextUserInterfaces.item_value — Function.

function item_value(item::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.jlgetch — Function.

function jlgetch(win::Union{Ptr{WINDOW},Nothing} = nothing)

Wait for an keystroke in the window win. If win is nothing, then getch() will be used instead of wgetch(win) to listen for the keystroke.

Returns

The raw value from the functions getch() or wgetch(win).

An instance of the structure Keystroke descrebing the keystroke.

TextUserInterfaces.keypad — Function.

function keypad(win::Ptr{WINDOW}, bf::Bool)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.menu_down_item — Method.

function menu_down_item(menu::TUI_MENU)

Move down to an item of the menu menu.

TextUserInterfaces.menu_down_line — Method.

function menu_down_line(menu::TUI_MENU)

Scroll down a line of the menu menu.

TextUserInterfaces.menu_down_page — Method.

function menu_down_page(menu::TUI_MENU)

Scroll down a page of the menu menu.

TextUserInterfaces.menu_driver — Function.

function menu_driver(menu::Ptr{Cvoid}, c::Integer)

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.menu_first_item — Method.

function menu_first_item(menu::TUI_MENU)

Move to the first item of the menu menu.

TextUserInterfaces.menu_last_item — Method.

function menu_last_item(menu::TUI_MENU)

Move to the last item of the menu menu.

TextUserInterfaces.menu_left_item — Method.

function menu_left_item(menu::TUI_MENU)

Move left to an item of the menu menu.

TextUserInterfaces.menu_next_item — Method.

function menu_next_item(menu::TUI_MENU)

Move to the next item of the menu menu.

TextUserInterfaces.menu_prev_item — Method.

function menu_prev_item(menu::TUI_MENU)

Move to the previous menu item of the menu.

TextUserInterfaces.menu_right_item — Method.

function menu_right_item(menu::TUI_MENU)

Move right to an item of the menu menu.

TextUserInterfaces.menu_toggle_item — Method.

function menu_toggle_item(menu::TUI_MENU)

Toggle the current item of the menu menu.

TextUserInterfaces.menu_up_item — Method.

function menu_up_item(menu::TUI_MENU)

Move up to an item of the menu menu.

TextUserInterfaces.menu_up_line — Method.

function menu_up_line(menu::TUI_MENU)

Scroll up a line of the menu menu.

TextUserInterfaces.menu_up_page — Method.

function menu_up_page(menu::TUI_MENU)

Scroll up a page of the menu menu.

TextUserInterfaces.mouseinterval — Function.

function mouseinterval(n::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.move_panel_to_top — Method.

function move_panel_to_top(panel::TUI_PANEL)

Move the panel panel to the top.

TextUserInterfaces.mvhline — Function.

function mvhline(y::Integer, x::Integer, ch::jlchtype, n::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.mvprintw — Function.

function mvprintw(y::Integer, x::Integer, str::T) where T<:AbstractString

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.mvvline — Function.

function mvvline(y::Integer, x::Integer, ch::jlchtype, n::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.mvwhline — Function.

function mvwhline(win::Ptr{WINDOW}, y::Integer, x::Integer, ch::jlchtype, n::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.mvwprintw — Function.

function mvwprintw(win::Ptr{WINDOW}, y::Integer, x::Integer, str::T) where T<:AbstractString

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.mvwvline — Function.

function mvwvline(win::Ptr{WINDOW}, y::Integer, x::Integer, ch::jlchtype, n::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.ncurses_color — Function.

function ncurses_color([foreground::Symbol, background::Symbol,] attrs::Int = 0; kwargs...)

Return a mask to apply a color format with the foreground color foreground, background color background, and the attributes attrs.

If the pair (foreground, background) is omitted, then the foreground and background color will not be changed.

Keywords

bold: If true, then apply bold format mask to attrs. (Default = false)

underline: If true, then apply underline format mask to attrs. (Default = false)

TextUserInterfaces.new_form — Function.

function new_form(fields::Vector{Ptr{Cvoid}})

Return type: Ptr{Nothing}

For more information, consult libform documentation.

TextUserInterfaces.new_item — Function.

function new_item(name::T, description::T) where T<:AbstractString

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.new_menu — Function.

function new_menu(items::Vector{Ptr{Cvoid}})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.new_panel — Function.

function new_panel(win::Ptr{WINDOW})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.newwin — Function.

function newwin(lines::Integer, cols::Integer, y::Integer, x::Integer)

Return type: Ptr{TextUserInterfaces.WINDOW}

For more information, consult libncurses documentation.

TextUserInterfaces.next_panel — Method.

function next_panel([panel::TUI_PANEL])

Make the next panel of panel be the top panel. If panel is not specified, then it will use the current top panel in the structure tui.

TextUserInterfaces.nodelay — Function.

function nodelay(win::Ptr{WINDOW}, bf::Bool)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.noecho — Function.

function noecho()

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.notimeout — Function.

function notimeout(win::Ptr{WINDOW}, bf::Bool)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.panel_userptr — Function.

function panel_userptr(pan::Ptr{Cvoid})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.pos_form_cursor — Function.

function pos_form_cursor(form::Ptr{Cvoid})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.pos_menu_cursor — Function.

function pos_menu_cursor(menu::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.post_form — Function.

function post_form(form::Ptr{Cvoid})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.post_menu — Function.

function post_menu(menu::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.post_menu — Method.

function post_menu(menu::TUI_MENU)

Post the menu menu.

TextUserInterfaces.prev_panel — Method.

function prev_panel([panel::TUI_PANEL])

Make the previous panel of panel be the top panel. If panel is not specified, then it will use the current top panel in the structure tui.

TextUserInterfaces.printw — Function.

function printw(str::T) where T<:AbstractString

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.refresh — Function.

function refresh()

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.refresh_all_windows — Method.

function refresh_all_windows()

Refresh all the windows, including the root window.

TextUserInterfaces.refresh_window — Method.

function refresh_window(id::String)

Refresh the window with id id and all its parents windows except for the root window.

TextUserInterfaces.refresh_window — Method.

function refresh_window(win::Ptr{WINDOW}; update = true)

Refresh the window win and all its parents windows except for the root window. If update is true, then doupdate() is called and the physical screen is updated.

TextUserInterfaces.selected_items — Method.

function selected_items(menu::TUI_MENU)

Return a Vector with the selected items in the menu menu.

TextUserInterfaces.selected_items_desc — Method.

function selected_items_desc(menu::TUI_MENU)

Return a Vector with the selected items descriptions in the menu menu.

TextUserInterfaces.selected_items_names — Method.

function selected_items_names(menu::TUI_MENU)

Return a Vector with the selected items names in the menu menu.

TextUserInterfaces.set_color — Method.

function set_color([win::TUI_WINDOW,] color::Int)

Set the color of the window win to color (see ncurses_color). If win is omitted, then it defaults to the root window.

TextUserInterfaces.set_field_back — Function.

function set_field_back(field::Ptr{Cvoid}, value::Integer)

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.set_field_buffer — Function.

function set_field_buffer(field::Ptr{Cvoid}, buf::Int, c::T) where T<:AbstractString

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.set_field_opts — Function.

function set_field_opts(field::Ptr{Cvoid}, form_options::Integer)

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.set_form_sub — Function.

function set_form_sub(form::Ptr{Cvoid}, win_form::Ptr{WINDOW})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.set_form_win — Function.

function set_form_win(form::Ptr{Cvoid}, win_form::Ptr{WINDOW})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.set_menu_format — Function.

function set_menu_format(menu::Ptr{Cvoid}, rows::Integer, cols::Integer)

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.set_menu_mark — Function.

function set_menu_mark(menu::Ptr{Cvoid}, mark::T) where T<:AbstractString

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.set_menu_opts — Function.

function set_menu_opts(menu::Ptr{Cvoid}, opts::Integer)

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.set_menu_sub — Function.

function set_menu_sub(menu::Ptr{Cvoid}, win::Ptr{WINDOW})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.set_menu_sub — Method.

function set_menu_sub(menu::TUI_MENU, win::TUI_WINDOW)

Set the menu menu sub-window to win.

TextUserInterfaces.set_menu_win — Function.

function set_menu_win(menu::Ptr{Cvoid}, win::Ptr{WINDOW})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.set_menu_win — Method.

function set_menu_win(menu::TUI_MENU, win::TUI_WINDOW)

Set menu menu window to win.

TextUserInterfaces.set_panel_userptr — Function.

function set_panel_userptr(pan::Ptr{Cvoid}, ptr::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.set_window_title! — Method.

function set_window_title!(win::TUI_WINDOW, title::AbstractString; ...)

Set the title of the window win to title.

Keywords

title_color: Color mask that will be used to print the title. See function ncurses_color. If negative, then the color will not be changed. (Default = -1)

TextUserInterfaces.show_panel — Function.

function show_panel(panel::Ptr{Cvoid})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.show_panel — Method.

function show_panel(panel::TUI_PANEL)

Show the panel panel.

TextUserInterfaces.start_color — Function.

function start_color()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.top_panel — Function.

function top_panel(pan::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.unpost_form — Function.

function unpost_form(form::Ptr{Cvoid})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.unpost_menu — Function.

function unpost_menu(menu::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.unpost_menu — Method.

function unpost_menu(menu::TUI_MENU)

Unpost the menu menu.

TextUserInterfaces.unset_color — Method.

function unset_color([win::TUI_WINDOW,] color::Number)

Unset the color color (see ncurses_color) in the window win. If win is omitted, then it defaults to the root window.

TextUserInterfaces.update_panels — Function.

function update_panels()

Return type: Nothing

For more information, consult libmenu documentation.

TextUserInterfaces.vline — Function.

function vline(ch::jlchtype, n::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.waddch — Function.

function waddch(win::Ptr{WINDOW}, ch::jlchtype)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.wattroff — Function.

function wattroff(win::Ptr{WINDOW}, attrs::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.wattron — Function.

function wattron(win::Ptr{WINDOW}, attrs::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.wborder — Function.

function wborder(win::Ptr{WINDOW}, ls::jlchtype, rs::jlchtype, ts::jlchtype, bs::jlchtype, tl::jlchtype, tr::jlchtype, bl::jlchtype, br::jlchtype)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.wborder — Method.

function wborder(win::Ptr{WINDOW})

Call the function wborder(win, 0, 0, 0, 0, 0, 0, 0, 0).

TextUserInterfaces.wclear — Function.

function wclear(win::Ptr{WINDOW})

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.wclrtobot — Function.

function wclrtobot(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.wclrtoeol — Function.

function wclrtoeol(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.werase — Function.

function werase(win::Ptr{WINDOW})

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.wgetch — Function.

function wgetch(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.whline — Function.

function whline(win::Ptr{WINDOW}, ch::jlchtype, n::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.window_print — Method.

function window_print(win::TUI_WINDOW, [row::Integer,] str::AbstractString; ...)

Print the string str at the window win in the row row. If the row is negative or omitted, then the current row will be used.

Keywords

alignment: Text alignemnt: :r for left, :c for center, and:lfor left. (**Default** =:l`)

pad: Padding to print the text. (Default = 0)

Remarks

If str has multiple lines, then all the lines will be aligned.

TextUserInterfaces.window_println — Method.

function window_println(win::TUI_WINDOW, [row::Integer,] str::AbstractString; ...)

Print the string str at the window win in the row row adding a break line character at the end. If the row is negative or omitted, then the current row will be used.

Keywords

alignment: Text alignemnt: :r for left, :c for center, and:lfor left. (**Default** =:l`)

pad: Padding to print the text. (Default = 0)

Remarks

If str has multiple lines, then all the lines will be aligned.

TextUserInterfaces.wmove — Function.

function wmove(win::Ptr{WINDOW}, y::Integer, x::Integer)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.wnoutrefresh — Function.

function wnoutrefresh(win::Ptr{WINDOW})

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.wprintw — Function.

function wprintw(win::Ptr{WINDOW}, str::T) where T<:AbstractString

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.wrefresh — Function.

function wrefresh(win::Ptr{WINDOW})

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.wtimeout — Function.

function wtimeout(win::Ptr{WINDOW}, delay::Integer)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.wvline — Function.

function wvline(win::Ptr{WINDOW}, ch::jlchtype, n::Integer)

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCURSES — Type.

struct NCURSES

This private structure handles some global variables that are used in the ncurses wrapper.

TextUserInterfaces.WINDOW — Type.

struct WINDOW

Handles a ncurses window.

TextUserInterfaces._get_window_cur_pos — Method.

function _get_window_cur_pos(win::TUI_WINDOW)

Get the cursor position of the window win and return it on a tuple (cur_y,cur_x). If the window is not initialized, then this function returns (-1,-1).

TextUserInterfaces._get_window_dims — Method.

function _get_window_dims(win::TUI_WINDOW)

Get the dimensions of the window win and return it on a tuple (dim_y,dim_x). If the window is not initialized, then this function returns (-1,-1).

TextUserInterfaces.@_ccallf — Macro.

macro _ccallf(expr)

Make a ccall to a libform function. The usage should be:

@_ccallf function(arg1::Type1, arg2::Type2, ...)::TypeReturn

It uses the global constant structure ncurses to call the function. Hence, it must be initialized.

TextUserInterfaces.@_ccallm — Macro.

macro _ccallm(expr)

Make a ccall to a libmenu function. The usage should be:

@_ccallf function(arg1::Type1, arg2::Type2, ...)::TypeReturn

It uses the global constant structure ncurses to call the function. Hence, it must be initialized.

TextUserInterfaces.@_ccalln — Macro.

macro _ccalln(expr)

Make a ccall to a libncurses function. The usage should be:

@_ccalln function(arg1::Type1, arg2::Type2, ...)::TypeReturn

It uses the global constant structure ncurses to call the function. Hence, it must be initialized.

TextUserInterfaces.@_ccallp — Macro.

macro _ccallm(expr)

Make a ccall to a libpanel function. The usage should be:

@_ccallf function(arg1::Type1, arg2::Type2, ...)::TypeReturn

It uses the global constant structure ncurses to call the function. Hence, it must be initialized.