NCurses library

Documentation for TextUserInterfaces.NCurses.

TextUserInterfaces.NCurses.fieldjust — Constant

Dictionary defining values of the field justification.

TextUserInterfaces.NCurses.fieldopts — Constant

Dictionary defining values of the field options.

TextUserInterfaces.NCurses.formcmd — Constant

Dictionary defining values of the form commands.

TextUserInterfaces.NCurses.menucmd — Constant

Dictionary defining values of the menu commands.

TextUserInterfaces.NCurses.NCURSES — Type

struct NCURSES

This private structure handles some global variables that are used in the ncurses wrapper.

TextUserInterfaces.NCurses.WINDOW — Type

struct WINDOW

Handles a ncurses window.

TextUserInterfaces.NCurses.ACS_ — Method

ACS_(s::Symbol)

Return the symbol s of the acs_map. For example, ACS_HLINE can be obtained from ACS_(:HLINE).

TextUserInterfaces.NCurses.COLS — Method

COLS()

Return the number of columns in the root window. It must be called after initscr().

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.LINES — Method

LINES()

Return the number of lines in the root window. It must be called after initscr().

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.TYPE_ALNUM — Method

TYPE_ALNUM()

Return a pointer to the global symbol TYPE_ALNUM of libform.

TextUserInterfaces.NCurses.TYPE_ALPHA — Method

TYPE_ALPHA()

Return a pointer to the global symbol TYPE_ALPHA of libform.

TextUserInterfaces.NCurses.TYPE_ENUM — Method

TYPE_ENUM()

Return a pointer to the global symbol TYPE_ENUM of libform.

TextUserInterfaces.NCurses.TYPE_INTEGER — Method

TYPE_INTEGER()

Return a pointer to the global symbol TYPE_INTEGER of libform.

TextUserInterfaces.NCurses.TYPE_IPV4 — Method

TYPE_IPV4()

Return a pointer to the global symbol TYPE_IPV4 of libform.

TextUserInterfaces.NCurses.TYPE_IPV6 — Method

TYPE_IPV6()

Return a pointer to the global symbol TYPE_IPV6 of libform.

TextUserInterfaces.NCurses.TYPE_NUMERIC — Method

TYPE_NUMERIC()

Return a pointer to the global symbol TYPE_NUMERIC of libform.

TextUserInterfaces.NCurses.TYPE_REGEXP — Method

TYPE_REGEXP()

Return a pointer to the global symbol TYPE_REGEXP of libform.

TextUserInterfaces.NCurses.attroff — Method

attroff(attrs::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.attron — Method

attron(attrs::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.bkgd — Method

bkgd(ch::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.bottom_panel — Method

bottom_panel(pan::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.box — Method

box(win::Ptr{WINDOW}, verch::T, horch::T) where T<:jlchtype

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.can_change_color — Method

can_change_color()

Return type: UInt8

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.cbreak — Method

cbreak()

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.clear — Method

clear()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.clrtobot — Method

clrtobot()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.clrtoeol — Method

clrtoeol()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.copywin — Method

copywin(scr::Ptr{WINDOW}, dest::Ptr{WINDOW}, sminrow::T, smincol::T, dminrow::T, dmincol::T, dmaxrow::T, dmaxcol::T, overlay::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.current_item — Method

function current_item(menu::Ptr{Cvoid})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.curs_set — Method

curs_set(visibility::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.curses_version — Method

curses_version()

Return the NCurses version in a named tuple with the following fields:

major: Major version.

minor: Minor version.

patch: Patch version.

TextUserInterfaces.NCurses.del_panel — Method

del_panel(panel::Ptr{Cvoid})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.delwin — Method

delwin(win::Ptr{WINDOW})

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.derwin — Method

derwin(win::Ptr{WINDOW}, nlines::T, ncols::T, begin_y::T, begin_x::T) where T<:Integer

Return type: Ptr{TextUserInterfaces.NCurses.WINDOW}

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.doupdate — Method

doupdate()

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.endwin — Method

endwin()

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.erase — Method

erase()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.field_buffer — Method

field_buffer(field::Ptr{Cvoid}, buffer::T) where T<:Integer

Return type: Cstring

For more information, consult libform documentation.

TextUserInterfaces.NCurses.form_driver — Method

form_driver(form::Ptr{Cvoid}, ch::T) where T<:Integer

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.free_field — Method

free_field(field::Ptr{Cvoid})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.free_form — Method

free_form(form::Ptr{Cvoid})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.free_item — Method

function free_item(item::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.free_menu — Method

function free_menu(menu::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.getbegx — Method

getbegx(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.getbegy — Method

getbegy(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.getch — Method

getch()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.getcurx — Method

getcurx(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.getcury — Method

getcury(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.getmaxx — Method

getmaxx(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.getmaxy — Method

getmaxy(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.has_colors — Method

has_colors()

Return type: UInt8

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.hide_panel — Method

hide_panel(panel::Ptr{Cvoid})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.hline — Method

hline(ch::Tc, n::Ti) where {Tc<:jlchtype,Ti<:Integer}

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.init_color — Method

init_color(color::T, r::T, g::T, b::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.init_pair — Method

init_pair(pair::T, f::T, b::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.initscr — Method

initscr()

Return type: Ptr{TextUserInterfaces.NCurses.WINDOW}

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.item_description — Method

function item_description(menu::Ptr{Cvoid})

Return type: Cstring

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.item_index — Method

function item_index(menu::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.item_name — Method

function item_name(menu::Ptr{Cvoid})

Return type: Cstring

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.item_value — Method

function item_value(item::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.keypad — Method

keypad(win::Ptr{WINDOW}, bf::Bool)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.leaveok — Method

leaveok(win::Ptr{WINDOW}, bf::Bool)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.load_ncurses — Method

load_ncurses([dir::String])

Load ncurses libraries at directory dir. If it is omitted or if it is empty, then the bundled Ncurses version in the package Ncurses_jll will be used.

TextUserInterfaces.NCurses.menu_driver — Method

function menu_driver(menu::Ptr{Cvoid}, c::T) where T<:Integer

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.mouseinterval — Method

mouseinterval(n::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.move_panel — Method

move_panel(panel::Ptr{Cvoid}, starty::T, startx::T) where T<:Integer

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.mvhline — Method

mvhline(y::Ti, x::Ti, ch::Tc, n::Ti) where {Tc<:jlchtype,Ti<:Integer}

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.mvprintw — Method

mvprintw(y::Ti, x::Ti, str::Ts) where {Ti<:Integer,Ts<:AbstractString}

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.mvvline — Method

mvvline(y::Ti, x::Ti, ch::Tc, n::Ti) where {Tc<:jlchtype,Ti<:Integer}

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.mvwhline — Method

mvwhline(win::Ptr{WINDOW}, y::Ti, x::Ti, ch::Tc, n::Ti) where {Tc<:jlchtype,Ti<:Integer}

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.mvwin — Method

mvwin(win::Ptr{WINDOW}, y::T, x::T) where T<:Integer

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.mvwprintw — Method

mvwprintw(win::Ptr{WINDOW}, y::Ti, x::Ti, str::Ts) where {Ti<:Integer,Ts<:AbstractString}

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.mvwvline — Method

mvwvline(win::Ptr{WINDOW}, y::Ti, x::Ti, ch::Tc, n::Ti) where {Tc<:jlchtype,Ti<:Integer}

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.new_form — Method

new_form(fields::Vector{Ptr{Cvoid}})

Return type: Ptr{Nothing}

For more information, consult libform documentation.

TextUserInterfaces.NCurses.new_item — Method

function new_item(name::T, description::T) where T<:AbstractString

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.new_menu — Method

function new_menu(items::Vector{Ptr{Cvoid}})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.new_panel — Method

new_panel(win::Ptr{WINDOW})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.newpad — Method

newpad(lines::T, cols::T) where T<:Integer

Return type: Ptr{TextUserInterfaces.NCurses.WINDOW}

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.newwin — Method

newwin(lines::T, cols::T, y::T, x::T) where T<:Integer

Return type: Ptr{TextUserInterfaces.NCurses.WINDOW}

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.nodelay — Method

nodelay(win::Ptr{WINDOW}, bf::Bool)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.noecho — Method

noecho()

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.notimeout — Method

notimeout(win::Ptr{WINDOW}, bf::Bool)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.panel_userptr — Method

panel_userptr(pan::Ptr{Cvoid})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.pnoutrefresh — Method

pnoutrefresh(win::Ptr{WINDOW}, pminrow::T, pmincol::T, sminrow::T, smincol::T, smaxrow::T, smaxcol::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.pos_form_cursor — Method

pos_form_cursor(form::Ptr{Cvoid})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.pos_menu_cursor — Method

function pos_menu_cursor(menu::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.post_form — Method

post_form(form::Ptr{Cvoid})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.post_menu — Method

function post_menu(menu::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.prefresh — Method

prefresh(win::Ptr{WINDOW}, pminrow::T, pmincol::T, sminrow::T, smincol::T, smaxrow::T, smaxcol::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.printw — Method

printw(str::T) where T<:AbstractString

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.refresh — Method

refresh()

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.resizeterm — Method

resizeterm(lines::T, columns::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.scrollok — Method

scrollok(win::Ptr{WINDOW}, bf::Bool)

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.set_field_back — Method

set_field_back(field::Ptr{Cvoid}, value::T) where T<:Integer

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_field_buffer — Method

set_field_buffer(field::Ptr{Cvoid}, buf::Ti, c::Ts) where {Ti<:Integer,Ts<:AbstractString}

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_field_just — Method

set_field_just(field::Ptr{Cvoid}, justification::T) where T<:Integer

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_field_opts — Method

set_field_opts(field::Ptr{Cvoid}, field_options::T) where T<:Integer

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_field_type — Method

set_field_type(field::Ptr{Cvoid}, type::Ptr{Cvoid}, regex::String)

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_field_type — Method

set_field_type(field::Ptr{Cvoid}, type::Ptr{Cvoid}, padding::T, vmin::Float64, vmax::Float64) where T<:Integer

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_field_type — Method

set_field_type(field::Ptr{Cvoid}, type::Ptr{Cvoid}, padding::T, vmin::T, vmax::T) where T<:Integer

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_field_type — Method

set_field_type(field::Ptr{Cvoid}, type::Ptr{Cvoid}, arg::T) where T<:Integer

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_field_type — Method

set_field_type(field::Ptr{Cvoid}, type::Ptr{Cvoid}, valuelist::Vector, checkcase::T, checkunique::T) where T<:Integer

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_form_opts — Method

set_form_opts(form::Ptr{Cvoid}, form_options::T) where T<:Integer

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_form_sub — Method

set_form_sub(form::Ptr{Cvoid}, win_form::Ptr{WINDOW})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_form_win — Method

set_form_win(form::Ptr{Cvoid}, win_form::Ptr{WINDOW})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.set_menu_format — Method

function set_menu_format(menu::Ptr{Cvoid}, rows::T, cols::T) where T<:Integer

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.set_menu_mark — Method

function set_menu_mark(menu::Ptr{Cvoid}, mark::T) where T<:AbstractString

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.set_menu_opts — Method

function set_menu_opts(menu::Ptr{Cvoid}, opts::T) where T<:Integer

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.set_menu_sub — Method

function set_menu_sub(menu::Ptr{Cvoid}, win::Ptr{WINDOW})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.set_menu_win — Method

function set_menu_win(menu::Ptr{Cvoid}, win::Ptr{WINDOW})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.set_panel_userptr — Method

set_panel_userptr(pan::Ptr{Cvoid}, ptr::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.show_panel — Method

show_panel(panel::Ptr{Cvoid})

Return type: Ptr{Nothing}

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.start_color — Method

start_color()

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.subpad — Method

subpad(win::Ptr{WINDOW}, nlines::T, ncols::T, begin_y::T, begin_x::T) where T<:Integer

Return type: Ptr{TextUserInterfaces.NCurses.WINDOW}

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.top_panel — Method

top_panel(pan::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.touchwin — Method

touchwin(win::Ptr{WINDOW})

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.unpost_form — Method

unpost_form(form::Ptr{Cvoid})

Return type: Int32

For more information, consult libform documentation.

TextUserInterfaces.NCurses.unpost_menu — Method

function unpost_menu(menu::Ptr{Cvoid})

Return type: Int32

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.update_panels — Method

update_panels()

Return type: Nothing

For more information, consult libmenu documentation.

TextUserInterfaces.NCurses.vline — Method

vline(ch::Tc, n::Ti) where {Tc<:jlchtype,Ti<:Integer}

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.waddch — Method

waddch(win::Ptr{WINDOW}, ch::T) where T<:jlchtype

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wattroff — Method

wattroff(win::Ptr{WINDOW}, attrs::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wattron — Method

wattron(win::Ptr{WINDOW}, attrs::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wbkgd — Method

wbkgd(win::Ptr{WINDOW}, ch::T) where T<:Integer

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wborder — Method

wborder(win::Ptr{WINDOW})

Call the function wborder(win, 0, 0, 0, 0, 0, 0, 0, 0).

TextUserInterfaces.NCurses.wborder — Method

wborder(win::Ptr{WINDOW}, ls::T, rs::T, ts::T, bs::T, tl::T, tr::T, bl::T, br::T) where T<:jlchtype

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wclear — Method

wclear(win::Ptr{WINDOW})

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wclrtobot — Method

wclrtobot(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wclrtoeol — Method

wclrtoeol(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.werase — Method

werase(win::Ptr{WINDOW})

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wgetch — Method

wgetch(win::Ptr{WINDOW})

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.whline — Method

whline(win::Ptr{WINDOW}, ch::Tc, n::Ti) where {Tc<:jlchtype,Ti<:Integer}

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wmove — Method

wmove(win::Ptr{WINDOW}, y::T, x::T) where T<:Integer

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wnoutrefresh — Method

wnoutrefresh(win::Ptr{WINDOW})

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wprintw — Method

wprintw(win::Ptr{WINDOW}, str::T) where T<:AbstractString

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wrefresh — Method

wrefresh(win::Ptr{WINDOW})

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wresize — Method

wresize(win::Ptr{WINDOW}, lines::T, cols::T) where T<:Integer

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wtimeout — Method

wtimeout(win::Ptr{WINDOW}, delay::T) where T<:Integer

Return type: Nothing

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.wvline — Method

wvline(win::Ptr{WINDOW}, ch::Tc, n::Ti) where {Tc<:jlchtype,Ti<:Integer}

Return type: Int32

For more information, consult libncurses documentation.

TextUserInterfaces.NCurses.@_ccallf — Macro

@_ccallf expr

Make a ccall to a libform function. The usage should be:

@_ccallf function(arg1::Type1, arg2::Type2, ...)::TypeReturn

It uses the global constant structure ncurses to call the function. Hence, it must be initialized.

TextUserInterfaces.NCurses.@_ccallm — Macro

@_ccallm expr

Make a ccall to a libmenu function. The usage should be:

@_ccallm function(arg1::Type1, arg2::Type2, ...)::TypeReturn

It uses the global constant structure ncurses to call the function. Hence, it must be initialized.

TextUserInterfaces.NCurses.@_ccalln — Macro

@_ccalln expr

Make a ccall to a libncurses function. The usage should be:

@_ccalln function(arg1::Type1, arg2::Type2, ...)::TypeReturn

It uses the global constant structure ncurses to call the function. Hence, it must be initialized.

TextUserInterfaces.NCurses.@_ccallp — Macro

@_ccallm expr

Make a ccall to a libpanel function. The usage should be:

@_ccallm function(arg1::Type1, arg2::Type2, ...)::TypeReturn

It uses the global constant structure ncurses to call the function. Hence, it must be initialized.