Library

Documentation for TextUserInterfaces.

TextUserInterfaces.Anchor — Type

struct Anchor

This structure defines an anchor to another object. It should be used in one of the fields anchor_* of ObjectPositioningConfiguration.

Fields

obj: Reference object.

side: Side of the reference object to which we will apply the attachment.

pad: A space between the anchors of the two objects.

Info

side is one of the following symbols: :left, :right, :top, :bottom, :center, or :middle.

TextUserInterfaces.Keystroke — Type

struct Keystorke

Structure that defines a keystroke.

Fields

value: String representing the keystroke.

ktype: Type of the key (:char, :F1, :up, etc.).

alt: true if ALT key was pressed (only valid if ktype != :char).

ctrl: true if CTRL key was pressed (only valid if ktype != :char).

shift: true if SHIFT key was pressed (only valid if ktype != :char).

TextUserInterfaces.ObjectPositioningConfiguration — Type

struct ObjectPositioningConfiguration

This structure defines the positioning of an object. It can also be created with the function newopc.

Fields

anchor_bottom: Bottom anchor specification. (Default = _no_anchor)

anchor_left: Left anchor specification. (Default = _no_anchor)

anchor_right: Right anchor specification. (Default = _no_anchor)

anchor_top: Top anchor specification. (Default = _no_anchor)

anchor_center: Center anchor specification. (Default = _no_anchor)

anchor_middle: Middle anchor specification. (Default = _no_anchor)

top: Absolute position of the object top side. (Default = -1)

left: Absolute position of the object left side. (Default = -1)

height: Height of the object. (Default = -1)

width: Width of the object. (Default = -1)

Info

If the absolute positioning arguments (top, left, height, and width) are negative, then it means that there is no information about them. Hence, the relative arguments (anchors) must provide the missing information.

TextUserInterfaces.newopc — Type

newopc(;kwargs...)

Create an instance of the structure ObjectPositioningConfiguration. The following keywords are available to specify the positioning:

anchor_bottom: Bottom anchor specification. (Default = _no_anchor)

anchor_left: Left anchor specification. (Default = _no_anchor)

anchor_right: Right anchor specification. (Default = _no_anchor)

anchor_top: Top anchor specification. (Default = _no_anchor)

anchor_center: Center anchor specification. (Default = _no_anchor)

anchor_middle: Middle anchor specification. (Default = _no_anchor)

top: Absolute position of the object top side. (Default = -1)

left: Absolute position of the object left side. (Default = -1)

height: Height of the object. (Default = -1)

width: Width of the object. (Default = -1)

Info

If the absolute positioning arguments (top, left, height, and width) are negative, then it means that there is no information about them. Hence, the relative arguments (anchors) must provide the missing information.

TextUserInterfaces._check_horizontal_anchor — Method

_check_horizontal_anchor(anchor::Anchor)

Check if the side parameter of anchor is valid for horizontal positioning. If anchor is _no_anchor, then true is always returned.

TextUserInterfaces._check_vertical_anchor — Method

_check_vertical_anchor(anchor::Nothing)
_check_vertical_anchor(anchor::Anchor)

Check if the side parameter of anchor is valid for vertical positioning. If anchor is _no_anchor, then true is always returned.

TextUserInterfaces._deselect_radio_button — Method

_select_radio_button(rb::WidgetRadioButton)

Deselect the radio button rb in its group name.

TextUserInterfaces._destroy_widget! — Method

_destroy_widget!(widget; refresh::Bool = true)

Private function that destroys a widget. This can be used if a new widget needs to reimplement the destroy function.

TextUserInterfaces._draw_title — Method

_draw_title(container::WidgetContainer)

Draw the title in the container container.

TextUserInterfaces._get_anchor — Method

_get_anchor(anchor::Anchor, parent)

Return the line or column related to the anchor anchor. If the object in anchor is the parent, then the positioning will be computed relative to the parent.

TextUserInterfaces._get_color_index — Method

_get_color_index(color::Symbol)

Return the index related to the color color.

TextUserInterfaces._get_window_cur_pos — Method

_get_window_cur_pos(win::Ptr{WINDOW})

Get the cursor position of the window win and return it on a tuple (cur_y,cur_x). If the window is not initialized, then this function returns (-1,-1).

TextUserInterfaces._get_window_dims — Method

_get_window_dims(win::Ptr{WINDOW})

Get the dimensions of the window win and return it on a tuple (dim_y,dim_x). If the window is not initialized, then this function returns (-1,-1).

TextUserInterfaces._next_widget — Method

_next_widget(container::WidgetContainer)

Move the focus of container container to the next widget.

TextUserInterfaces._previous_widget — Method

_previous_widget(container::WidgetContainer)

Move the focus of container container to the previous widget.

TextUserInterfaces._process_horizontal_info — Method

_process_horizontal_info(opc::ObjectPositioningConfiguration)

Process the horizontal positioning information in opc and write the variable horizontal of the same structure. The possible horizontal positioning information are:

:aleft_aright: Left and right anchors were specified.

:aleft_height: Left anchor and height were specified.

:aright_height: Right anchor and height were specified.

:acenter_height: Center anchor and height were specified.

:right_height: Right and height were specified.

:unknown: Insufficient information to compute the horizontal positioning.

TextUserInterfaces._process_positioning_value — Method

_process_positioning_value(v, dim::Symbol, parent::WidgetParent)

Process the positioning value v related to the dimension dim of the parent widget parent. dim can be :height or :width.

If v is an Int, then it return v.

TextUserInterfaces._process_vertical_info — Method

_process_vertical_info(opc::ObjectPositioningConfiguration)

Process the vertical positioning information in opc and write the variable vertical of the same structure. The possible vertical positioning information are:

:abottom_atop: Bottom and top anchors were specified.

:abottom_height: Bottom anchor and height were specified.

:atop_height: Top anchor and height were specified.

:amiddle_height: Middle anchor and height were specified.

:top_height: Top and height were specified.

:unknown: Insufficient information to compute the vertical positioning.

TextUserInterfaces._select_radio_button — Method

_select_radio_button(rb::WidgetRadioButton)

Select the radio button rb in its group name.

TextUserInterfaces._str — Method

_str(wpc::ObjectPositioningConfiguration)

Convert the information in wpc to a string for debugging purposes.

TextUserInterfaces.accept_focus — Method

accept_focus(widget)

Return true is the widget widget accepts focus or false otherwise.

TextUserInterfaces.accept_focus — Method

accept_focus(window::Window)

Check if the window window can accept focus and, if it can, then perform the actions to change the focus.

TextUserInterfaces.add_widget! — Method

add_widget!(container::WidgetContainer, widget::Widget)

Add the widget widget to the container `container.

TextUserInterfaces.add_widget! — Method

add_widget!(win::Window, widget::Widget)

Add widget widget to the window win. If the win already have a widget, then it will be replaced.

TextUserInterfaces.app_main_loop — Method

app_main_loop(;fprev::Function = (k)->true, fpost::Function = (k)->true, destroy_on_exit::Bool = true, manage_window_switch::Bool = true)

Initialize the application main loop.

The function fprev is called after every pressed key and before the focus is processed. It must have the following signature:

fprev(k::Keystroke)::Bool

If it returns true, then the keystroke is processed. Otherwise, the software waits for another keystroke.

The function fpost is called after the keystroke is processed. It must have the following signature:

fpost(k::Keystroke)::Bool

If it return true, then the loop continues and a next keysotrke is requested. Otherwise, the loop is terminated.

Configuration keywords

destroy_on_exit: If true, then destroy_tui() is called when we exit from the loop.

manage_window_switch: If true, then F2 and F3 move to next and previous windows.

TextUserInterfaces.change_label — Method

change_label(button::WidgetButton, label::AbstractString)

Change the label of button button to label.

TextUserInterfaces.change_text — Method

change_text(widget::WidgetANSILabel, new_text::AbstractString; alignment = :l)

Change to text of the label widget widget to new_text.

The text alignment in the widget can be selected by the keyword alignment, which can be:

:l: left alignment (default);

:c: Center alignment; or

:r: Right alignment.

TextUserInterfaces.change_text — Method

change_text(widget::WidgetLabel, new_text::AbstractString; alignment = :l, color::Int = -1)

Change to text of the label widget widget to new_text.

The text alignment in the widget can be selected by the keyword alignment, which can be:

:l: left alignment (default);

:c: Center alignment; or

:r: Right alignment.

The text color can be selected by the keyword color. It it is negative (default), then the current color will not be changed.

TextUserInterfaces.change_value — Method

change_value(widget::WidgetProgressBar, new_value::Int; color::Int = -1)

Change the value of the progress bar to new_value.

The color can be selected by the keyword color. It it is negative (default), then the current color will not be changed.

TextUserInterfaces.clear_data! — Method

clear_data!(widget::WidgetForm)

Clear the data in all the input fields in the form widget.

TextUserInterfaces.clear_data! — Method

clear_data!(widget::WidgetInputField)

Clear the data in the input field widget.

TextUserInterfaces.compute_object_positioning — Method

compute_object_positioning(opc::ObjectPositioningConfiguration, parent)

Compute the object position based on the configuration opc and on its parent object parent.

Return

The object height.

The object width.

The top position w.r.t. the parent object.

The left position w.r.t. the parent object.

TextUserInterfaces.create_widget — Function

create_widget(T, opc::ObjectPositioningConfiguration, args...; kwargs...)
create_widget(T, parent::WidgetParent, opc::ObjectPositioningConfiguration, args...; kwargs)

Create the widget of type T using the positioning configuration opc.

Additional arguments and keywords related to each widget can be passed using args and kwargs, respectively.

If the second signature is called, then the created widget is added to the parent parent.

TextUserInterfaces.create_window — Function

function create_window(opc::ObjectPositioningConfiguration = newopc(), id::String = "";  kwargs...)

Create a window. The window ID id is used to identify the new window in the global window list. The size and location of the window is configured by the object opc.

Keyword

bcols: Number of columns in the window buffer. This will be automatically increased to, at least, fit the viewable part of the window. (Default = 0)

blines: Number of lines in the window buffer. This will be automatically increased to, at least, fit the viewable part of the window. (Default = 0)

border: If true, then the window will have a border. (Default = true)

border_color: Color mask that will be used to print the border. See function ncurses_color. If negative, then the color will not be changed. (Default = -1)

focusable: If true, then the window can have focus. Otherwise, all focus request will be rejected. (Default = true)

title: The title of the window, which will only be printed if border is true. (Default = "")

title_color: Color mask that will be used to print the title. See function ncurses_color. If negative, then the color will not be changed. (Default = -1)

TextUserInterfaces.create_window_with_container — Method

create_window_with_container(vargs...; kwargs...)

Create a window with a container as its widget. The arguments and keywords are the same ones of the function create_window. The container will have the same size of the window buffer.

Return

The created window.

The created container.

TextUserInterfaces.destroy_all_windows — Method

destroy_all_windows()

Destroy all windows managed by the TUI.

TextUserInterfaces.destroy_tui — Method

destroy_tui()

Destroy the Text User Interface (TUI).

TextUserInterfaces.destroy_widget! — Method

destroy_widget!(widget; refresh::Bool = true)

Destroy the widget widget.

If refresh is true (default), then a full refresh will be performed on the parent window. Otherwise, no refresh will be performed.

TextUserInterfaces.destroy_widget_buffer! — Method

destroy_widget_buffer!(widget::Widget)

Destroy the buffer of the widget widget.

TextUserInterfaces.destroy_window! — Method

destroy_window!(win::Window)

Destroy the window win.

TextUserInterfaces.get_buffer — Method

get_buffer(widget)

Return the buffer of the widget widget.

TextUserInterfaces.get_buffer — Method

get_buffer(win::Window)

Return the buffer of the window win.

TextUserInterfaces.get_color_pair — Method

get_color_pair(foreground::Int, background::Int)

Return the ID of the color pair (foreground, background), or nothing if the color pair is not initialized.

TextUserInterfaces.get_current_item — Method

get_selected_item(widget::WidgetListBox)

Return the ID of the current item of the list box widget and the data associated with it.

TextUserInterfaces.get_data — Method

get_data(widget::WidgetInputField)

Return a vector with the data of all fields.

TextUserInterfaces.get_data — Method

get_data(widget::WidgetInputField)

Get the data of widget as a string. If the validator is enabled and the data is not valid, then return nothing.

TextUserInterfaces.get_data — Method

get_data(widget::WidgetListBox)

Return the data of the list box widget.

TextUserInterfaces.get_focused_window — Method

get_focused_window()

Return the focused window.

TextUserInterfaces.get_height — Function

get_height(object)

Return the usable height of object object.

TextUserInterfaces.get_height_for_child — Function

get_height_for_child

Return the usable height of the object for a child object.

TextUserInterfaces.get_left — Function

get_bottom(object)

Return the left of the object w.r.t. its parent.

TextUserInterfaces.get_left_for_child — Function

get_left_for_child

Return the left of the object for a child object.

TextUserInterfaces.get_parent — Method

get_parent(widget)

Return the parent of the widget widget.

TextUserInterfaces.get_parent — Method

get_parent(win::Window)

Return nothing since the window has no parent.

TextUserInterfaces.get_selected — Method

get_selected(group_name::AbstractString)

Return the WidgetRadioButton that is selected in group with name group_name. If the group_name does not exists or if no button is selected, then nothing is returned.

TextUserInterfaces.get_selected — Method

get_selected(widget::WidgetListBox)

Return an array of Bool indicating which elements of the list box widget are selected.

TextUserInterfaces.get_top — Function

get_top(object)

Return the top of the object w.r.t. its parent.

TextUserInterfaces.get_top_for_child — Function

get_top_for_child

Return the top of the object for a child object.

TextUserInterfaces.get_width — Function

get_width(object)

Return the usable width of object object.

TextUserInterfaces.get_width_for_child — Function

get_width_for_child

Return the usable width of object object for a child object.

TextUserInterfaces.has_focus — Method

has_focus(container::WidgetContainer, widget)

Return true if the widget widget is in focus on container container, or false otherwise.

TextUserInterfaces.has_focus — Method

has_focus(window::Window, widget)

Return true if the widget widget is in focus on window window, or false otherwise.

TextUserInterfaces.hide_window — Method

hide_window(win::Window)

Hide the window win.

TextUserInterfaces.init_color_pair — Method

init_color_pair(foreground::Symbol, background::Symbol)

Initialize the color pair (foreground, background) and return its ID. If the pair already exists, then just the function just returns its ID.

TextUserInterfaces.init_focus_manager — Method

init_focus_manager()

Initialization of the focus manager. The elements in focus_chain are iterated to find the first one that can accept the focus.

TextUserInterfaces.init_tui — Function

init_tui(dir::String = "")

Initialize the Text User Interface (TUI). The full-path of the ncurses directory can be specified by dir. If it is empty or omitted, then it will look on the default library directories.

TextUserInterfaces.init_widget_buffer! — Method

init_widget_buffer!(widget::Widget)

Initialize the buffer of widget widget. The variables opc and parent must be set before calling this function. If the buffer is already initialized, then it will be deleted first.

TextUserInterfaces.jlgetch — Function

jlgetch(win::Union{Ptr{WINDOW},Nothing} = nothing)

Wait for an keystroke in the window win and return it (see Keystroke). If win is nothing, then getch() will be used instead of wgetch(win) to listen for the keystroke.

TextUserInterfaces.log_message — Function

log_message(level::Int, msg::AbstractString, id::AbstractString = "")

Log the message msg with level level. The ID of the called can be specified by id.

If a line is @log_pad X, then the following lines will have a padding of X.

TextUserInterfaces.modify_color — Method

modify_color([name::Symbol, ]id::Int, r::Int, g::Int, b::Int)

Modify the color ID id to the RGB value (r,g,b). If the symbol name is available, then the user can select this color ID by using name instead of the id.

If the color name name already exists, then nothing will be changed.

Notice that the range for the RGB color components is [0,1000].

If the color was initialized, then it returns the color ID. Otherwise, it returns -1.

TextUserInterfaces.move_view — Method

move_view(win::Window, y::Int, x::Int; update::Bool = true)

Move the origin of the view of window win to the position (y,x). This routine makes sure that the view will never reach positions outside the buffer.

TextUserInterfaces.move_view_inc — Method

move_view_inc(win::Window; Δy::Int, Δx::Int; kwargs...)

Move the view of the window win to the position (y+Δy, x+Δx). This function has the same set of keywords of the function move_view.

TextUserInterfaces.move_window — Method

move_window(win::Window, starty::Int, startx::Int)

Move the window win to the position (starty, startx).

TextUserInterfaces.move_window_to_top — Method

move_window_to_top(win::Window)

Move window win to the top.

TextUserInterfaces.ncurses_color — Function

ncurses_color([foreground, background,] attrs::Integer = 0; kwargs...)

Return a mask to apply a color format with the foreground color foreground, background color background, and the attributes attrs.

If the pair (foreground, background) is omitted, then the foreground and background color will not be changed.

The colors can be specified by their names using Symbol or by their indices using Int.

Keywords

bold: If true, then apply bold format mask to attrs. (Default = false)

underline: If true, then apply underline format mask to attrs. (Default = false)

TextUserInterfaces.next_window — Method

next_window()

Move the focus to the next window.

TextUserInterfaces.obj_desc — Method

obj_desc(obj)

Return a string with the description of the object obj formed by:

<Object type> (<Object address if mutable>)

TextUserInterfaces.obj_to_ptr — Method

 function obj_to_ptr(obj)

Returns the hexadecimal representation of the address of the object obj. It only works with mutable objects. If obj is immutable, then 0x0 will be returned.

TextUserInterfaces.previous_window — Method

previous_window()

Move the focus to the previous window.

TextUserInterfaces.process_focus — Method

process_focus(widget, k::Keystroke)

Process the actions when widget widget is in focus and the keystroke k is pressed. If it returns false, then it is means that the widget was not capable to process the focus. Otherwise, it must return true.

TextUserInterfaces.process_focus — Method

process_focus(k::Keystroke)

Process the focus considering the user's keystorke k.

TextUserInterfaces.process_focus — Method

process_focus(window::Window, k::Keystroke)

Process the focus on window window due to keystroke k.

TextUserInterfaces.redraw — Function

redraw(widget)

Redraw the widget inside its content window cwin.

TextUserInterfaces.refresh_all_windows — Method

refresh_all_windows()

Refresh all the windows, including the root window.

TextUserInterfaces.refresh_window — Method

refresh_window(id::String)

Refresh the window with id id and all its parents windows except for the root window.

TextUserInterfaces.refresh_window — Method

refresh_window(container::WidgetContainer; force_redraw::Bool = false)

Ask the parent widget to refresh the window. If force_redraw is true, then all widgets in the window will be updated.

TextUserInterfaces.refresh_window — Method

refresh_window(win::Window; force_redraw = false)

Refresh the window win and its widget. If the view needs to be updated (see view_needs_update) or if force_redraw is true, then the content of the buffer will be copied to the view before updating.

TextUserInterfaces.release_focus — Method

release_focus(widget)

Request focus to be released. It should return true if the focus can be released or false otherwise.

TextUserInterfaces.remove_widget! — Method

remove_widget!(container::WidgetContainer, widget::Widget)

Remove the widget widget from the container container.

TextUserInterfaces.remove_widget! — Method

remove_widget!(win::Window, widget::Widget)

Remove the widget widget from the window win. If widget does not belong to win, then nothing is done.

TextUserInterfaces.reposition! — Function

reposition!(object)

Reposition the object baesd on the stored configuration.

TextUserInterfaces.request_focus — Method

request_focus(widget)

Request to focus to the widget widget.

TextUserInterfaces.request_focus — Method

request_focus(container::WidgetContainer, widget)

Request the focus to the widget widget of the container container. It returns true if the focus could be changed or false otherwise.

TextUserInterfaces.request_focus — Method

request_focus(win::Window)

Request the focus to the window win. If win cannot get the focus, then nothing happens and it returns false. If wincan get the focus, then the focus is passed to it and the function returns true.

Remarks

Even if win is in the focus chain, the focus_id will not change by requesting focus to win. This means that the window focus order is not altered by this function.

TextUserInterfaces.request_next_widget — Method

request_next_widget(container::WidgetContainer)

Request the next widget in container. It returns true if a widget has get the focus or false otherwise.

TextUserInterfaces.request_prev_widget — Method

request_prev_widget(container::WidgetContainer)

Request the previous widget in container. It returns true if a widget has get the focus or false otherwise.

TextUserInterfaces.request_update — Method

request_update(widget)

Request update of the widget widget.

TextUserInterfaces.request_update — Method

request_update(win::Window)

Request update of the window win because its widget was updated.

TextUserInterfaces.request_view_update — Method

request_view_update(win::Window)

Request to update the view of window win. Notice that this must also request update on all parent windows until the root window.

TextUserInterfaces.require_cursor — Method

require_cursor()

If true, then the physical cursor will be shown and the position will be updated according to its position in the widget window. Otherwise, the physical cursor will be hidden.

TextUserInterfaces.select_item — Method

select_item(widget::WidgetListBox, id::Int)

Select the item id in the list box widget. Notice that id refers to the position of the item in the array data.

TextUserInterfaces.set_color — Method

set_color([win::Window,] color::Int)

Set the color of the window win to color (see ncurses_color). If win is omitted, then it defaults to the root window.

TextUserInterfaces.set_focus_chain — Method

set_focus_chain(wins::Window...; new_focus_id::Int = 1)

Set the focus chain, i.e. the ordered list of windows that can receive the focus. The keyword new_focus_id can be set to specify which element is currently focused in the new chain.

TextUserInterfaces.set_widget_color — Method

set_widget_color(widget, color::Int)

Set the background color of widget to color.

TextUserInterfaces.set_window_title — Method

set_window_title(win::Window, title::AbstractString; ...)

Set the title of the window win to title.

Keywords

title_color: Color mask that will be used to print the title. See function ncurses_color. If negative, then the color will not be changed. (Default = -1)

TextUserInterfaces.show_window — Method

show_window(win::Window)

Show the window win.

TextUserInterfaces.sync_cursor — Method

sync_cursor(widget::WidgetContainer)

Synchronize the cursor to the position of the focused widget in container container. This is necessary because all the operations are done in the buffer and then copied to the view.

TextUserInterfaces.sync_cursor — Method

sync_cursor(window::Window)

Synchronize the cursor to the position of the focused widget in window window. This is necessary because all the operations are done in the buffer and then copied to the view.

TextUserInterfaces.tui_update — Method

tui_update()

Update the entire TUI.

TextUserInterfaces.unset_color — Method

unset_color([win::Window,] color::Number)

Unset the color color (see ncurses_color) in the window win. If win is omitted, then it defaults to the root window.

TextUserInterfaces.update — Method

update(widget; force_redraw = false)

Update the widget by calling the function redraw. This function returns true if the widget needed to be updated of false otherwise.

If force_redraw is true, then the widget will be updated even if it is not needed.

TextUserInterfaces.update_view — Method

update_view(win::Window; force::Bool = false)

Update the view of window win by copying the contents from the buffer. If the view does not need to be updated (see view_needs_update), then nothing is done. If the keyword force is true, then the copy will always happen.

Return

It returns true if the view has been updated and false otherwise.

TextUserInterfaces.validate_str — Method

validate_str(str::AbstractString, v)

Validate the string str using the validator v. v is an element of the type that will be used to validate or a regex.

TextUserInterfaces.@composed_widget — Macro

@composed_widget(ex)

Declare a structure of a composed widget.

TextUserInterfaces.@connect_signal — Macro

@connect_signal(obj, signal::Symbol, f::Symbol, vargs...)

Connect the signal signal of the object obj to the function f passing the additional arguments vargs. Thus, when signal is emitted by obj, then fcall will be executed.

TextUserInterfaces.@derived_widget — Macro

@derived_widget(ex)

Declare a structure of a derived widget.

TextUserInterfaces.@disconnect_signal — Macro

@disconnect_signal(obj, signal::Symbol)

Disconnect the signal signal from object obj.

TextUserInterfaces.@emit_signal — Macro

@emit_signal(obj, signal::Symbol, params...)

Emit the signal signal of the object obj with the parameters params..., causing to execute the connected function.

TextUserInterfaces.@forward_signal — Macro

@forward_signal(src, dest, signal::Symbol)

Forward the signal signal from src to dest. This means that every time that the signal signal is generated in src, then the function in dest will be called.

TextUserInterfaces.@log — Macro

macro log(level, msg, id)

Log the messagem msg with level level of the caller id

TextUserInterfaces.@log — Macro

macro log(level, msg)

Log the message msg with level level.

TextUserInterfaces.@signal — Macro

@signal(name)

Create the signal named name. This must be used inside the widget structure that must be declared with @with_kw option (see package Parameters.jl).

TextUserInterfaces.@widget — Macro

@widget(ex)

Declare a structure of a widget.